Finish day 20.
I don't want to be bothered with all of the tile flipping
and rotating and piecing together work that needs to be done,
so as with day 22 from last year, I'm going to download someone
else's solution and run it. This time i'm going to do it inside
a docker container.